Look through DelegatingMethodHandle.getTarget() during inlining
Recognize DelegatingMethodHandle.getTarget() and determine its result as
a known object in:

- InterpreterEmulator, which can now find the call target of a
  later invokeBasic() call whose receiver is the result; and

- Method handle transformer, which can now refine such an invokeBasic()
  call to match the target found by InterpreterEmulator.

With this, the compiler can inline the DelegatingMethodHandle target.

Because DelegatingMethodHandle.getTarget() is abstract, the target must
be determined based on the concrete type of the DelegatingMethodHandle.
So far only java/lang/invoke/MethodHandleImpl$CountingWrapper has been
observed.

Co-authored-by: Daryl Maier <maier@ca.ibm.com>