RCP: Restore immortal classloaders using persisted state
When using RCP to restore an application where modules are used,
a new J9ClassLoader is created and assigned to the java app
class loader. This causes applications to fail as it introduces
a mismatch between the J9ClassLoader in the persisted state and
the J9ClassLoader in the restore run. The fix is to initialize
the application class loader upon construction like boot and
platform class loaders and assign it to the persisted J9ClassLoader
in the RCP cache. The changes are to implement JNI interfaces
VM.isRestoreRun() and VM.rcpAssignClassLoader() to initialize the
builtin class loaders by
OutOfLineINL_com_ibm_oti_vm_VM_rcpAssignClassLoader() for RCP restore
run. A new constant J9_CLASSLOADER_TYPE_APP is added to identify the
application class loader.

Fixes: #22287

Co-authored-by: Tobi Ajila <tobi_ajila@ca.ibm.com>
Co-authored-by: Babneet Singh <sbabneet@ca.ibm.com>