RAM Class Persistence: Recover protection domain
During the training run, the protection domain of J9Classes is
discarded. Without the protection domain of J9Class, programs
like Open liberty servers will fail and it's required by Java
security. The change is to recover the protection domain when
the classes are loaded in the production run. The classes loaded
in defineClassCommon may have a protection domain object
associated with them. In a restore run it is required that these
classes are loaded in the same manner. Given that hashClassTableAt
is used as a fast path check for class loading in many places, it
must be enforced that classes which are frozen and have JCL define
class flag are only returned when they are being looked up from
defineClassesCommon.

Fixes: #22257