Add a flag to the functions of class hash table
To simplify the changes when the result from the class hash table
is conditional on flags of J9Class, a flag is added to the
functions hashClassTableAt() and hashClassTableAtString() to filter
the result from hash class table. The changes will save a number
of changes all over of the source codes affected and can be reused
by other purpose similar to this case. It also makes the update of
source codes easy if the conditions are changed in the future.
The PR is required by PR #22537

Co-authored-by: Tobi Ajila tobi_ajila@ca.ibm.com