Restrict use of higher allowance threshold during ECS
At veryHot and scorching compilations, use a higher analyzed size
threshold for methods that are always worth inlining along with
their sub call graphs. This ensures that we do not prematurely
terminate ECS along call graphs that should be inlined. At lower
opt-levels, due to the the smaller inlining budget available,
benefitting from analyzing such call graphs with a larger size
threshold is rare. Therefore, contributing signficiantly to the
overall compilation overhead with enabling this at warm and hot
opt-levels is not justifiable based on current observations in
various workloads.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>