Fix rare case where we attempt to patch a helper call
Since JDK11 (JEP 181, nestmates) "invokevirtual" can be used to call a
private method. This results in _virtualUnresolvedHelper() snippets
being used indefinitely exposing us to the possibility that the helper
will call a method that was just recompiled and patched to call
_samplingPatchCallSite(). Since the helper is reached using a "bl"
instruction the patching logic will think that the call site can be
patched to call the new recompiled body, but that's not the case. This
fix will return false from ppcCodePatching() when the bl being patched
is calling outside the code-cache or is calling through a trampoline to
a helper.