RAM Class Persistence: Fix Open liberty server hangs
The Open liberty server hangs on startup when loading classes from
RCP cache and initializes the frozen classes. It is caused by
deadlock. Two threads wait for each other on RCP cache mutex and
class table mutex in classes loading from RCP cache in
loadNonArrayClass(). The changes are to pass the flag
J9_FINDCLASS_FLAG_EXISTING_ONLY to class table lookup functions.
If it's set for a frozen class, then it won't be loaded from RCP
cache. The real problem is that a thread requested a class with
J9_FINDCLASS_FLAG_EXISTING_ONLY is attempting to perform a RCP
classload (loadWarmClass) which should never happen. We would have
similar deadlock issues in a non-RCP scenario if a thread
requesting a class with J9_FINDCLASS_FLAG_EXISTING_ONLY attempted
to do a full classload. This is why the only threads that are
allowed to load classes are VM threads and the JIT threads always
call with J9_FINDCLASS_FLAG_EXISTING_ONLY.

Fixes: #22753

Co-authored-by: Tobi Ajila tobi_ajila@ca.ibm.com