Remove breakOnDFSet functionality
There was logic inserted into JITed code and the JIT runtime many years ago to
aid in investigating a peculiar bug. As I recall, the issue was that on x86
macOS, the CPU direction flag was being flipped on some calls to the C runtime
which interfered with the operation of JIT inlined arraycopy. It was not
possible to predict when or why this flag modification was occurring, so logic
was introduced across certain entry points to the C runtime to detect a
direction flag toggle when it occurred. We did not suspect a problem with the
OpenJ9 Java VM, but either a C runtime or macOS issue.

This problem has not been seen in over five years despite exhaustive testing on
x86 macOS in the meantime. The problem was repeatable every few days if I recall
correctly, so I think it has either gone into deep remission or has been fixed.

Nevertheless, this diagnostic logic is just technical debt now and is removed
by this PR.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>